+++ manage/src/main/java/com/system/manage/repositories/professorRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.system.manage.models.professor.Professor;
 
-public interface professorRepository extends JpaRepository<Professor,Integer> {
+public interface professorRepository extends JpaRepository<Professor,Long> {
     
 }
