+++ manage/src/main/java/com/system/manage/repositories/turmaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.system.manage.models.turma.Turma;
 
-public interface turmaRepository extends JpaRepository<Turma,Integer> {
+public interface turmaRepository extends JpaRepository<Turma, Long> {
     
 }
